+++ src/converse-core.js
@@ -151,6 +151,7 @@
         Strophe.addNamespace('ROSTERX', 'http://jabber.org/protocol/rosterx');
         Strophe.addNamespace('XFORM', 'jabber:x:data');
         Strophe.addNamespace('NICK', 'http://jabber.org/protocol/nick');
+        Strophe.addNamespace('HINTS', 'urn:xmpp:hints');
 
         // Instance level constants
         this.TIMEOUTS = { // Set as module attr so that we can override in tests.