+++ docs/CHANGES.md
@@ -8,6 +8,7 @@
 - Chat bot messages don't appear when they have the same ids as their commands.
   [jcbrand]
 - Updated onDisconnected method to fire disconnected event even if auto_reconnect is false. [kamranzafar]
+- #553 Add processing hints to OTR messages [jcbrand]
 
 ## 1.0.2 (2016-05-24)